requirements.txt
pygetwindow
pillow
pyautogui
opencv-python
rich
pydantic

#### Development Related ####
ipykernel
ipywidgets
pre-commit
pytest
sh
